refactor: factor out import cycle in aiguard (#19616)
The `ddtrace.aiguard._api_client ↔ ddtrace.aiguard._redaction` cycle is fixed via Pattern 1 (extract shared type):

  - New `ddtrace/aiguard/_types.py` holds the Message, ContentPart, ImageURL, ToolCall, Function TypedDicts that both
  modules need.
  - `_api_client.py` now imports these from `_types.py` (and still re-exports them, so `ddtrace/aiguard/__init__.py`'s lazy
  `getattr(_api_client, name)` re-export mechanism keeps working unchanged).
  - `_redaction.py` imports Message directly from `_types.py` — no more `TYPE_CHECKING`-only import back to `_api_client`.

(cherry picked from commit 88eb8a3ce6bae719c7ca8d41019b318583ee9e0a)
